Aktivitas Validation Update
Update autorisasi penggunaan contrroller aktivitas
Update kemunculan tombol-tombol

+++ controllers/AktivitasController.php
@@ -44,9 +44,14 @@ class AktivitasController extends Controller
         $searchModel = new AktivitasSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 		
-		
 		$model=new Aktivitas();
-		$data=$model->getAll();
+		if(Yii::$app->user->identity->jabatan == 'Coordinator'){
+			$data=$model->getAllCoor(Yii::$app->user->identity->nik);
+		}
+		else{
+			$data=$model->getAll();
+		}
+		
         return $this->render('index', [
             'data' => $data
         ]);
@@ -73,13 +78,20 @@ class AktivitasController extends Controller
 	/**
      * Creates a new Aktivitas model.
      * If creation is successful, the browser will be redirected to the 'view' page.
+	 * Hanya untuk supervisor / coordinator / administrator(temp)
      * @return mixed
      */
     public function actionCreate()
     {
+    	$jabatan=Yii::$app->user->identity->jabatan;
     	if (\Yii::$app->user->isGuest) {
             return $this->redirect('/propensi/web');
         }
+		
+		if(!($jabatan=='Project Manager' || $jabatan=='Supervisor' || $jabatan='Administrator')){
+			return $this->redirect('/propensi/web/index.php/aktivitas');
+		}
+			
         $model = new Aktivitas();
 		$modelSite=Site::find()->all();
 
@@ -96,15 +108,28 @@ class AktivitasController extends Controller
     /**
      * Updates an existing Aktivitas model.
      * If update is successful, the browser will be redirected to the 'view' page.
+	 ** Hanya untuk supervisor / coordinator / administrator(temp)
+	 * Hanya ketika aktivitas belum di approve
      * @param integer $id
      * @return mixed
      */
     public function actionUpdate($id)
     {
+    	$jabatan=Yii::$app->user->identity->jabatan;
+		
     	if (\Yii::$app->user->isGuest) {
             return $this->redirect('/propensi/web');
         }
+		
+		if(!($jabatan=='Project Manager' || $jabatan=='Supervisor' || $jabatan='Administrator')){
+			return $this->redirect('/propensi/web/index.php/aktivitas');
+		}
+		
         $model = $this->findModel($id);
+		
+		if($model->status_approval_supervi=='1' || $model->status_approval_pm=='1'){
+			return $this->redirect('/propensi/web/index.php/aktivitas');
+		}
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->id]);
@@ -118,14 +143,28 @@ class AktivitasController extends Controller
     /**
      * Deletes an existing Aktivitas model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
+	 * Hanya untuk supervisor / coordinator / administrator(temp)
+	 * Hanya ketika belum diapprove 
      * @param integer $id
      * @return mixed
      */
     public function actionDelete($id)
     {
+    	$jabatan=Yii::$app->user->identity->jabatan;
     	if (\Yii::$app->user->isGuest) {
             return $this->redirect('/propensi/web');
         }
+		
+		if(!($jabatan=='Project Manager' || $jabatan=='Supervisor' || $jabatan='Administrator')){
+			return $this->redirect('/propensi/web/index.php/aktivitas');
+		}
+		
+		$model = $this->findModel($id);
+		
+		if($model->status_approval_supervi=='1' || $model->status_approval_pm=='1'){
+			return $this->redirect('/propensi/web/index.php/aktivitas');
+		}	
+		
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -134,12 +173,17 @@ class AktivitasController extends Controller
 	/**
 	 * Approval Method
 	 * Untuk melakukan approval dengan diberikan notes di keterangan!
+	 * Hanya bisa dilakukan oleh supervisor / project manager / admin
 	 */
 	 public function actionApprove($id){
 	 	if (\Yii::$app->user->isGuest) {
             return $this->redirect('/propensi/web');
         }
 		
+		if(!($jabatan=='Project Manager' || $jabatan=='Supervisor' || $jabatan='Administrator')){
+			return $this->redirect('/propensi/web/index.php/aktivitas');
+		}
+						
 		$model=$this->findModel($id);
 		if(!Yii::$app->user->identity->jabatan=='Supervisor' || !Yii::$app->user->identity->jabatan=='Project Manager')
 		{
@@ -151,8 +195,20 @@ class AktivitasController extends Controller
 	 
 	 /**
 	  * Approval Process Method
+	  * Hanya untuk Supervisor dan Project Manager
 	  */
 	  public function actionApproveprocess(){
+			
+	  	$jabatan=Yii::$app->user->identity->jabatan;
+		
+		if(!($jabatan=='Supervisor' || $jabatan='Project Manager')){
+			return $this->redirect('/propensi/web/index.php/aktivitas');
+		}
+
+		if (\Yii::$app->user->isGuest) {
+            return $this->redirect('/propensi/web');
+        }		
+		
 	  	 $data=array();	
 	  	 $data['id']=$_GET['id'];
 	  	 $data['user']=$_GET['user'];
@@ -160,6 +216,7 @@ class AktivitasController extends Controller
 		 $data['notes']=$_GET['notes'];
 		 $data['keterangan']=$_GET['keterangan'];
 		 $data['username']=$_GET['username'];
+		 $data['statusApproval']=$_GET['approveSP'];
 		 
 		 $model=new Aktivitas();
 		 
@@ -176,9 +233,18 @@ class AktivitasController extends Controller
 				
 		 //Approval PM
 		 if($data['user']=='Project Manager'){
-		 	
+		 	if($data['statusApproval']!=1){
+		 		$this->redirect("/propensi/web/index.php/aktivitas/approve?id=$data[id]&status=failApprove");
+		 	}
+			else if($model->approve($data)=='1'){
+				if($data['status']=='approve')
+		 			$this->redirect("/propensi/web/index.php/aktivitas/approve?id=$data[id]&status=suksesApprove");
+		 		if($data['status']=='reject')
+					$this->redirect("/propensi/web/index.php/aktivitas/approve?id=$data[id]&status=suksesReject");
+				}	
+			}
 		 }
-	  } 
+	  
 
     /**
      * Finds the Aktivitas model based on its primary key value.
@@ -187,7 +253,7 @@ class AktivitasController extends Controller
      * @return Aktivitas the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
-    protected function findModel($id)
+   protected function findModel($id)
     {
     	if (\Yii::$app->user->isGuest) {
             return $this->redirect('/propensi/web');
